Full file configuration implemented for LP, only comments in code missint to be ready to publish

+++ Configuration.h
@@ -13,8 +13,8 @@
 class Configuration {
 public:
 	byte mac[6];// = {0x00, 0x12, 0xFB, 0x95, 0x59, 0xCF};
-	unsigned int mqttServerIP[4]; // = {192, 168, 1, 190};
-	int mqttServerPort = 1883;
+	uint8_t mqttServerIP[4]; // = {192, 168, 1, 190};
+	uint16_t mqttServerPort = 1883;
 	bool isDebug = true;
 	char* instanceName;
